[Sukumar] Display bottom sheet on tapping markers on the map

+++ lib/models/bid_item.dart
@@ -5,7 +5,7 @@ class BidItem {
 
   String name;
 
-  double availableQuantity;
+  int availableQuantity;
   double specifiedPRice;
   double bidQuantity;
   double bidPrice;
